refactor: :recycle: update imports, add _document file
