+++ components/date/index.ts
@@ -1,0 +1,1 @@
+export { Date } from './Date'
