+++ components/layout/index.ts
@@ -1,0 +1,2 @@
+export { Layout } from './layout';
+export { siteTitle } from './layout';
